Add a retrying background batch producer
I need this for a project of mine, and it seemed
like it belonged in this library rather than in
my application code.

The batch producer:

* Has a single goroutine running in the background
* Uses a channel for buffering and backpressure
* Flushes batches either when a specified batch size
  is in the buffer or when an optional specified
  interval occurs
* Accepts a StatReceiver and will regularly send
  stats to it
* Will retry when HTTP errors occur or are received
* Will retry individual records that fail

The retry behavior in particular is important because
Kinesis sometimes returns 500 responses to valid
requests, and this is considered normal by AWS:
https://forums.aws.amazon.com/thread.jspa?messageID=531831

For more details see the godoc comments.